+++ ACG/lab4/src/point/Point.cpp
@@ -1,7 +1,7 @@
 #include "Point.h"
 
 
-void Point::set(int x, int y, int z) {
+void Point::set(double x, double y, double z) {
     this->x = x;
     this->y = y;
     this->z = z;
